Clean up locale and move type narrowing function to avoid circular dependency
