--- frontend/src/scripts/modules/locale/locale.de.ts
@@ -1,63 +1,0 @@
-import { I18nKey } from "./locale.en";
-
-export const TEXT_DE = {
-    ai: "KI",
-    chooseMode: "Spielmodus wählen",
-    confirm_password: "Passwort bestätigen",
-    create_lobby: "Lobby erstellen",
-    create_tournament: "Ein Turnier erstellen",
-    difficulty: "Schwierigkeit",
-    display_name: "Anzeigename",
-    easy: "Einfach",
-    enter_names: "Spielernamen eingeben:",
-    failed_generate_chart: "Fehler beim Generieren des Diagramms.",
-    failed_query: "Fehler beim Abrufen der Benutzerdaten.",
-    game: "Spiel",
-    game_id: "Spiel-ID",
-    games_played: "Gespielte Spiele",
-    hard: "Schwer",
-    home: "Startseite",
-    initialize_controller: "Spielcontroller nicht initialisiert.",
-    join_lobby: "Lobby beitreten",
-    lang: "Sprache: DE",
-    leaderboard: "Bestenliste",
-    local: "Lokal",
-    login: "Einloggen",
-    login_failed: "Anmeldung fehlgeschlagen. Bitte überprüfen Sie Ihre Anmeldeinformationen.",
-    logout: "Abmelden",
-    medium: "Mittel",
-    name_player: "Benenne Spieler ",
-    not_enough_data: "Nicht genügend Daten, um das Diagramm zu generieren.",
-    online: "Online",
-    passw_not_match: "Passwörter stimmen nicht überein.",
-    password: "Passwort",
-    player_names_required: "Bitte geben Sie Namen für alle Spieler ein. Keine Duplikate erlaubt.",
-    player_number: "Spieleranzahl",
-    please_enter_name: "Bitte geben Sie einen Namen für Spieler",
-    play_ai: "Gegen KI spielen",
-    profile: "Profil",
-    profile_picture: "Spieler-Profilbild",
-    quickplay: "Schnelles Spiel",
-    rank: "Rang",
-    register: "Registrieren",
-    register_failed: "Registrierung fehlgeschlagen. Bitte versuchen Sie es erneut.",
-    setup: "Einrichtung",
-    setup_ai: "KI-Modus",
-    setup_choose_mode: "Spielmodus wählen",
-    setup_local: "Lokal",
-    setup_online: "Online spielen",
-    setup_play: "Spielen",
-    setup_play_local: "Lokal spielen",
-    setup_tournament_mode: "Turniermodus",
-    select_Difficulty: "Wähle eine Schwierigkeit.",
-    select_player_amount: "Bitte wählen Sie eine Anzahl von Spielern aus.",
-    start_tournament: "Turnier starten",
-    title: "zv-Transzendenz",
-    tournament_mode: "Turniermodus",
-    TOTP: "TOTP",
-    username: "Nutzername",
-    wins: "Siege",
-    your_profile: "Dein Profil",
-    your_stats: "Deine Statistiken:",
-    code_default: "Translation Needed", // Needs attention to replace, for type reasons
-} satisfies Record<I18nKey, string>;
