--- frontend/src/scripts/modules/locale/locale.en.ts
@@ -1,65 +1,0 @@
-export const TEXT_EN = {
-    ai: "AI",
-    chooseMode: "Choose Game Mode",
-    confirm_password: "Confirm Password",
-    create_lobby: "Create Lobby",
-    create_tournament: "Create a tournament",
-    difficulty: "Difficulty",
-    display_name: "Display Name",
-    easy: "Easy",
-    enter_names: "Enter player names:",
-    failed_generate_chart: "Failed to generate chart.",
-    failed_query: "Failed to query user data.",
-    game: "Game",
-    game_id: "Game ID",
-    games_played: "Games Played",
-    hard: "Hard",
-    home: "Home",
-    initialize_controller: "Game controller not initialized.",
-    join_lobby: "Join Lobby",
-    lang: "Language: EN",
-    leaderboard: "Leaderboard",
-    local: "Local",
-    login: "Login",
-    login_failed: "Login failed. Please check your credentials.",
-    logout: "Logout",
-    medium: "Medium",
-    name_player: "Name Player ",
-    not_enough_data: "Not enough data to generate chart.",
-    online: "Online",
-    passw_not_match: "Passwords do not match.",
-    password: "Password",
-    player_names_required: "Please enter names for all players. No duplicates allowed.",
-    player_number: "Player Number",
-    please_enter_name: "Please enter a name for Player",
-    play_ai: "Play AI",
-    profile: "Profile",
-    profile_picture: "Player Profile Picture",
-    quickplay: "Quickplay",
-    rank: "Rank",
-    register: "Register",
-    register_failed: "Registration failed. Please try again.",
-    setup: "Setup",
-    setup_ai: "AI-Mode",
-    setup_choose_mode: "Choose Game Mode", // duplicate
-    setup_local: "Local", // duplicate
-    setup_online: "Play Online",
-    setup_play: "Play",
-    setup_play_local: "Play Local",
-    setup_tournament_mode: "Tournament Mode",
-    select_Difficulty: "Please select a difficulty.",
-    select_player_amount: "Please select an amount of players.",
-    start_tournament: "Start Tournament",
-    title: "ft-transcendence",
-    tournament_mode: "Tournament Mode",
-    TOTP: "TOTP",
-    username: "Username",
-    wins: "Wins",
-    your_profile: "Your Profile",
-    your_stats: "Your Stats:",
-    code_default: "Translation Needed", // Needs attention to replace, for type reasons
-} as const;
-
-// English translation will be the only source of truth, all other languages must have same keys
-export type I18nKey = keyof typeof TEXT_EN;
-export const TEXT_KEYS = Object.keys(TEXT_EN) as I18nKey[];
