--- frontend/src/scripts/modules/locale/locale.es.ts
@@ -1,64 +1,0 @@
-import { I18nKey } from "./locale.en";
-
-export const TEXT_ES = {
-    ai: "IA",
-    chooseMode: "Elegir modo de juego",
-    confirm_password: "Confirmar contraseña",
-    create_lobby: "Crear sala",
-    create_tournament: "Crear un torneo",
-    difficulty: "Dificultad",
-    display_name: "Nombre para mostrar",
-    easy: "Fácil",
-    enter_names: "Ingrese los nombres de los jugadores:",
-    failed_generate_chart: "Error al generar el gráfico.",
-    failed_query: "Error al consultar los datos del usuario.",
-    game: "Juego",
-    game_id: "ID del juego",
-    games_played: "Juegos jugados",
-    hard: "Difícil",
-    home: "Inicio",
-    initialize_controller: "Controlador de juego no inicializado.",
-    join_lobby: "Unirse a la sala",
-    lang: "Idioma: ES",
-    leaderboard: "Clasificación",
-    local: "Local",
-    login: "Iniciar sesión",
-    login_failed: "Error al iniciar sesión. Verifique sus credenciales.",
-    logout: "Cerrar sesión",
-    medium: "Medio",
-    name_player: "Nombre del jugador ",
-    not_enough_data: "No hay suficientes datos para generar el gráfico.",
-    online: "En línea",
-    passw_not_match: "Las contraseñas no coinciden.",
-    password: "Contraseña",
-    player_names_required:
-        "Por favor, ingrese los nombres de todos los jugadores. No se permiten duplicados.",
-    player_number: "Número de jugadores",
-    please_enter_name: "Por favor, ingrese un nombre para el jugador",
-    play_ai: "Jugar contra IA",
-    profile: "Perfil",
-    profile_picture: "Foto de perfil del jugador",
-    quickplay: "Juego rápido",
-    rank: "Rango",
-    register: "Registrarse",
-    register_failed: "Error al registrarse. Por favor, inténtelo de nuevo.",
-    setup: "Configuración",
-    setup_ai: "IA-Modo",
-    setup_choose_mode: "Elegir modo de juego",
-    setup_local: "Local",
-    setup_online: "Jugar en línea",
-    setup_play: "Jugar",
-    setup_play_local: "Jugar localmente",
-    setup_tournament_mode: "Modo torneo",
-    select_Difficulty: "Por favor, seleccione una dificultad.",
-    select_player_amount: "Por favor, seleccione un número de jugadores.",
-    start_tournament: "Iniciar torneo",
-    title: "cd-transcendencia",
-    tournament_mode: "Modo torneo",
-    TOTP: "TOTP",
-    username: "Nombre de usuario",
-    wins: "Victorias",
-    your_profile: "Tu perfil",
-    your_stats: "Tus estadísticas:",
-    code_default: "Translation Needed", // Needs attention to replace, for type reasons
-} satisfies Record<I18nKey, string>;
