--- frontend/src/scripts/modules/locale/locale.fr.ts
@@ -1,63 +1,0 @@
-import { I18nKey } from "./locale.en";
-
-export const TEXT_FR = {
-    ai: "IA",
-    chooseMode: "Choisir le mode de jeu",
-    confirm_password: "Confirmer le mot de passe",
-    create_lobby: "Créer un salon",
-    create_tournament: "Créer un tournoi",
-    difficulty: "Difficulté",
-    display_name: "Nom d'affichage",
-    easy: "Facile",
-    enter_names: "Entrez les noms des joueurs :",
-    failed_generate_chart: "Échec de la génération du graphique.",
-    failed_query: "Échec de la requête des données utilisateur.",
-    game: "Jeu",
-    game_id: "ID de jeu",
-    games_played: "Parties jouées",
-    hard: "Difficile",
-    home: "Accueil",
-    initialize_controller: "Contrôleur de jeu non initialisé.",
-    join_lobby: "Rejoindre un salon",
-    lang: "Langue : FR",
-    leaderboard: "Classement",
-    local: "Local",
-    login: "Connexion",
-    login_failed: "Échec de la connexion. Veuillez vérifier vos identifiants.",
-    logout: "Déconnexion",
-    medium: "Moyen",
-    name_player: "Nom du joueur ",
-    not_enough_data: "Pas assez de données pour générer le graphique.",
-    online: "En ligne",
-    passw_not_match: "Les mots de passe ne correspondent pas.",
-    password: "Mot de passe",
-    player_names_required: "Veuillez entrer les noms de tous les joueurs, sans doublons.",
-    player_number: "Nombre de joueurs",
-    please_enter_name: "Veuillez entrer un nom pour le joueur",
-    play_ai: "Jouer contre l'IA",
-    profile: "Profil",
-    profile_picture: "Photo de profil du joueur",
-    quickplay: "Jeu rapide",
-    rank: "Rang",
-    register: "S'inscrire",
-    register_failed: "Échec de l'inscription. Veuillez réessayer.",
-    setup: "Configuration",
-    setup_ai: "IA-Mode",
-    setup_choose_mode: "Choisir le mode de jeu",
-    setup_local: "Local",
-    setup_online: "Jouer en ligne",
-    setup_play: "Jouer",
-    setup_play_local: "Jouer en local",
-    setup_tournament_mode: "Mode tournoi",
-    select_Difficulty: "Veuillez sélectionner une difficulté.",
-    select_player_amount: "Veuillez sélectionner un nombre de joueurs.",
-    start_tournament: "Démarrer le tournoi",
-    title: "qd-transcendance",
-    tournament_mode: "Mode tournoi",
-    TOTP: "TOTP",
-    username: "Nom d'utilisateur",
-    wins: "Victoires",
-    your_profile: "Votre profil",
-    your_stats: "Vos statistiques :",
-    code_default: "Translation Needed", // Needs attention to replace, for type reasons
-} satisfies Record<I18nKey, string>;
