--- frontend/src/scripts/modules/locale/locale.text.ts
@@ -1,47 +1,0 @@
-import { LanguageOpts } from "./locale.store";
-
-// Refactor like this later maybe
-// All caps as code to not accidentally set up something as translatable
-export const TEXT = {
-    AI: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    CHOOSE_GAME_MODE: { en: "Choose Game Mode", de: "", fr: "", es: "" },
-    CONFIRM_PASSWORD: { en: "Confirm Password", de: "", fr: "", es: "" },
-    CREATE_LOBBY: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    CREATE_TOURNAMENT: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    DIFFICULTY: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    DISPLAY_NAME: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    EASY: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    ENTER_NAMES: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    FAILED_GENERATE_CHARD: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    FAILED_QUERY: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    GAME: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    GAME_ID: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    GAMES_PLAYED: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    HARD: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    HOME: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    JOIN_LOBBY: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    LANG: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    LEADERBOARD: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    LOCAL: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    LOGIN: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    LOGIN_FAILED: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    LOGOUT: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    MEDIUM: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    NAME_PLAYER: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    NOT_ENOUGH_DATA: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    ONLINE: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    PASSWORD_NOT_MATCH: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    PASSWORD: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    PLAYERS_NAMES_REQUIRED: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    PLAYER_NUMBER: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    PLAYER_ENTER_NAME: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    PLAY_AI: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    PROFILE: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    PROFILE_PICTURE: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    QUICKPLAY: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    RANK: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    REGISTER: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    REGISTER_FAILED: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    SETUP: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-    SETUP_AI: { en: "AI", de: "KI", fr: "IA", es: "IA" },
-} as const satisfies Record<string, Record<LanguageOpts, string>>; //
