+++ frontend/src/scripts/ui/components/Status.ts
@@ -1,6 +1,6 @@
 import { twMerge } from "tailwind-merge";
-import { I18nKey } from "../../modules/locale/locale.en";
-import { isValidKey } from "../../modules/locale/locale.store";
+import { I18nKey } from "../../modules/locale/locale.translation";
+import { isValidKey } from "../../modules/locale/locale.utils";
 import { getText } from "../../modules/locale/locale.utils";
 import { createEl } from "../../utils/dom-helper";
 import { createParagraph } from "./Paragraph";
